+++ CHANGELOG.md
@@ -58,3 +58,4 @@
 - Pasted File From Folder to group
 # Version 0.0.29
 - add $mgp, $mgr ở Mobile
+- add definition controller